Separated io server code
