+++ server/index.js
@@ -1,16 +1,18 @@
 const http = require('http') // Prod should use https
-const fs = require('fs')
 const express = require('express')
 const consola = require('consola')
 const { Nuxt, Builder } = require('nuxt')
+const config = require('../nuxt.config.js')
+const { IOServer } = require('./io')
 
 // Boiler-plate
 const app = express()
 const server = http.createServer(app) // Prod should separate server code from client and use https
-const io = require('socket.io')(server)
+const altServer = http.createServer()
+altServer.listen(4000, 'localhost')
+const ioServer = IOServer(altServer)
 
 // Import and Set Nuxt.js options
-const config = require('../nuxt.config.js')
 config.dev = process.env.NODE_ENV !== 'production'
 
 async function start() {
@@ -35,29 +37,9 @@ async function start() {
     message: `Server listening on http://${host}:${port}`,
     badge: true
   })
-
-  const ioChannels = fs
-    .readdirSync('./server/channels')
-    .map((f) => f.replace('.js', ''))
-
-  ioChannels.forEach((channel) => {
-    io.of(`/${channel}`).on('connection', (socket) => {
-      consola.info('socket.io client connected to', channel)
-      const svc = require(`./channels/${channel}`).Svc()
-      Object.entries(svc).forEach(([evt, fn]) => {
-        if (typeof fn === 'function') {
-          socket.on(evt, (msg, cb) => {
-            const { notifyEvt = 'progress' } = msg
-            fn({
-              notify: (data) => {
-                socket.emit(notifyEvt, data)
-              },
-              ...msg
-            }).then(cb)
-          })
-        }
-      })
-    })
-  })
 }
-start()
+
+if (require.main === module) {
+  start()
+  ioServer.start()
+}
